fix(k8s_mysql): check whether clients installed
